Struct massa_network_worker::network_worker::NetworkWorker

pub struct NetworkWorker {

    cfg: NetworkConfig,
    pub(crate) keypair: KeyPair,
    pub(crate) self_node_id: NodeId,
    listener: Listener,
    establisher: Establisher,
    pub(crate) peer_info_db: PeerInfoDatabase,
    controller_command_rx: Receiver<NetworkCommand>,
    controller_manager_rx: Receiver<NetworkManagementCommand>,
    pub(crate) running_handshakes: HashSet<ConnectionId>,
    handshake_futures: FuturesUnordered<JoinHandle<(ConnectionId, Result<(NodeId, ReadBinder, WriteBinder), NetworkError>)>>,
    handshake_peer_list_futures: FuturesUnordered<JoinHandle<()>>,
    node_event_rx: Receiver<NodeEvent>,
    pub(crate) active_nodes: HashMap<NodeId, (ConnectionId, Sender<NodeCommand>)>,
    node_worker_handles: FuturesUnordered<JoinHandle<(NodeId, Result<ConnectionClosureReason, NetworkError>)>>,
    pub(crate) active_connections: HashMap<ConnectionId, (IpAddr, bool)>,
    version: Version,
    pub(crate) event: EventSender,

}

Real job is done by network worker

Fields
cfg: NetworkConfig

Network configuration.

keypair: KeyPair

Our keypair.

self_node_id: NodeId

Our node id.

listener: Listener

Listener part of the establisher.

establisher: Establisher

The connection establisher.

peer_info_db: PeerInfoDatabase

Database with peer information.

controller_command_rx: Receiver<NetworkCommand>

Receiver for network commands

controller_manager_rx: Receiver<NetworkManagementCommand>

Receiver for network management commands

running_handshakes: HashSet<ConnectionId>

Set of connection id of node with running handshake.

handshake_futures: FuturesUnordered<JoinHandle<(ConnectionId, Result<(NodeId, ReadBinder, WriteBinder), NetworkError>)>>

Running handshakes futures.

handshake_peer_list_futures: FuturesUnordered<JoinHandle<()>>

Running handshakes that send a list of peers.

node_event_rx: Receiver<NodeEvent>

Receiving channel for node events.

active_nodes: HashMap<NodeId, (ConnectionId, Sender<NodeCommand>)>

Ids of active nodes mapped to Connection id, node command sender and handle on the associated node worker.

node_worker_handles: FuturesUnordered<JoinHandle<(NodeId, Result<ConnectionClosureReason, NetworkError>)>>

Node worker handles

active_connections: HashMap<ConnectionId, (IpAddr, bool)>

Map of connection to ip, is_outgoing.

version: Version

Node version

event: EventSender

Event sender

Implementations

impl NetworkWorker

pub fn new( cfg: NetworkConfig, keypair: KeyPair, listener: Listener, establisher: Establisher, peer_info_db: PeerInfoDatabase, _: NetworkWorkerChannels, version: Version ) -> NetworkWorker

Creates a new NetworkWorker

Arguments

cfg: Network configuration.

listener: Listener part of the establisher.

establisher: The connection establisher.

peer_info_db: Database with peer information.

controller_command_rx: Channel receiving network commands.

controller_event_tx: Channel sending out network events.

controller_manager_rx: Channel receiving network management commands.

pub async fn run_loop(self) -> Result<(), NetworkError>

Runs the main loop of the network worker There is a tokio::select! inside the loop

async fn on_handshake_finished( &mut self, new_connection_id: ConnectionId, outcome: Result<(NodeId, ReadBinder, WriteBinder), NetworkError> ) -> Result<(), NetworkError>

Manages finished handshakes. Only used by the worker.

Arguments

new_connection_id: connection id of the connection that should be established here.

outcome: result returned by a handshake.

async fn connection_closed( &mut self, id: ConnectionId, reason: ConnectionClosureReason ) -> Result<(), NetworkError>

async fn manage_network_command( &mut self, cmd: NetworkCommand ) -> Result<(), NetworkError>

Manages network commands Only used inside worker’s run_loop

Arguments

cmd : command to process.

peer_info_db: Database with peer information.

active_connections: hashmap linking connection id to ÌpAddr to whether connection is outgoing (true)

event_tx: channel to send network events out.

Command implementation

Some of the commands are just forwarded to the NodeWorker that manage the real connection between nodes. Some other commands has an impact on the current worker.

Whatever the behavior of the command, we better have to look at network_cmd_impl.rs where the commands are implemented.

ex: NetworkCommand::AskForBlocks => on_ask_bfor_block_cmd(...)

async fn manage_out_connections( &mut self, res: Result<(ReadHalf, WriteHalf)>, ip_addr: IpAddr, cur_connection_id: &mut ConnectionId ) -> Result<(), NetworkError>

Manages out connection Only used inside worker’s run_loop

Arguments

res: (reader, writer) in a result coming out of out_connecting_futures

ip_addr: distant address we are trying to reach.

cur_connection_id: connection id of the node we are trying to reach

async fn manage_in_connections( &mut self, res: Result<(ReadHalf, WriteHalf, SocketAddr)>, cur_connection_id: &mut ConnectionId ) -> Result<(), NetworkError>

Manages in connection Only used inside worker’s run_loop

Try a connection with an incoming node, if success insert the remote address at the index connection_id inside self.active_connections and call self.manage_successful_connection

If the connection failed with MaxPeersConnectionReached, mock the handshake and send a list of advertisable peer ips.

Arguments

re : (reader, writer, socketAddr) in a result coming out of the listener

cur_connection_id: connection id of the node we are trying to reach

fn try_send_peer_list_in_handshake( &self, reader: ReadHalf, writer: WriteHalf, remote_addr: SocketAddr )

Start to mock a handshake and try to send a message with a list of peers. The function is used while manage_in_connections() if the current node reach the number of maximum in connection.

Timeout if the connection as not been done really quickly resolved but don’t throw the timeout error.

In connection node            Curr node
       |                          |
       |------------------------->|      : Try to connect but the
       |                          |        curr node reach the max
       |                          |        number of node.
       |                          |      : Connection success anyway
       |                          |        and the in connection enter
       |<------------------------>|        in `HandshakeWorker::run()`
       |  symmetric read & write   |

In the symmetric read & write the current node simulate a handshake managed by the connection node in HandshakeWorker::run(), the current node send a ListPeer as a message.

Spawn a future in self.handshake_peer_list_futures managed by the main loop.

fn manage_successful_connection( &mut self, connection_id: ConnectionId, reader: ReadHalf, writer: WriteHalf ) -> Result<(), NetworkError>

Manage a successful incoming and outgoing connection, Check if we’re not already running an handshake for connection_id by inserting the connection id in self.running_handshakes Add a new handshake to perform in self.handshake_futures to be handle in the main loop.

Return an handshake error if connection already running/waiting

async fn on_node_event(&mut self, evt: NodeEvent) -> Result<(), NetworkError>

Manages node events. Only used by the worker.

Argument

evt: optional node event to process.