Function massa_network_worker::peer_info_database::cleanup_peers

pub(crate) fn cleanup_peers(
    cfg: &NetworkConfig,
    peers: &mut HashMap<IpAddr, PeerInfo>,
    opt_new_peers: Option<&Vec<IpAddr>>,
    ban_timeout: MassaTime
) -> Result<(), NetworkError>

Cleans up the peer database using max values provided by NetworkConfig.ProtocolConfig. If opt_new_peers is provided, adds its contents as well.

Note: only standard non-active peers are counted when clipping to size limits.

Arguments :

cfg: NetworkConfig

peers: peers to clean up

opt_new_peers: optional peers to add to the database

clock_compensation: to be sync with server time

ban_timeout: after that time we forget we banned a peer